fix: fix for tests GitHub

+++ main-service/src/main/java/ru/practicum/main/models/Event.java
@@ -25,7 +25,7 @@ public class Event {
     @OneToOne
     @JoinColumn(name = "category_id", referencedColumnName = "id")
     private Category category;
-    private int confirmedRequests;
+    private Long confirmedRequests;
     @Column(name = "created_On")
     private LocalDateTime createdOn;
     @Column(length = 7000)
